creating arp package with relevant data structures
